Fixed incompatibilities with Vaadin 7.2

+++ src/main/java/org/vaadin/easyuploads/client/ui/VMultiUpload.java
@@ -264,7 +264,7 @@ public class VMultiUpload extends SimplePanel implements Paintable {
 	 * @return
 	 */
 	private String serialize(VHtml5File file) {
-		int size = file.getSize();
+		int size = (int) file.getSize();
 		String name = file.getName();
 		String type = file.getType();
 		return size + DELIM + name + DELIM + type + DELIM;
